Feature - Consoslidate Incrementally
For very large sparse locations (500K+ files in a location), regular consolidation mostly fails because of S3 issues. Keeping the location as is is also not an option ideally. In Consolidate Incrementally we only emit required X (default to 50K) files from the PathLister Implementation. This way the Pail will try to create splits only for these 50K files and consolidate them. After the consolidation the smaller 50K file would be deleted and the process can be repeated.