Making the consolidateIncremental truely incremental
In our earlier approach we used to get all the files, but return only a subset of it for consolidation. For some extreme cases the number of files to fetch from S3 itself is too large that the Consolidator fails with OOM. This change now introduces a set of changes where we always get only Integer.MAX_FILES (as default behaviour). Override them for consolidateIncrementally.